Add memcache-client comparisons.
This benchmark starts to show how GC becomes a huge issue in Ruby. Just
loading Rubygems causes remix-stash to take over a 45% hit on my machine.
This is because most of the time is spent GCing not executing on 1.8. 1.9
does slightly better but still gets at least a 40% hit on large payloads.